fix all tests
