+++ packages/trackerUserInfoLambda/.jest/setEnvVars.js
@@ -1,7 +1,17 @@
-process.env.idpTokenPath = "https://dummytoken.com/token"
-process.env.oidcjwksEndpoint = "https://dummyauth.com/.well-known/jwks.json"
-process.env.UserPoolIdentityProvider = "DummyPoolIdentityProvider"
-process.env.oidcClientId = "valid_aud"
-process.env.oidcIssuer = "valid_iss"
 process.env.TokenMappingTableName = "dummyTable"
-process.env.userInfoEndpoint = "https://dummyauth.com/userinfo"
+
+process.env.CIS2_OIDC_ISSUER = "valid_cis2_iss"
+process.env.CIS2_OIDC_CLIENT_ID = "valid_cis2_aud"
+process.env.CIS2_OIDC_HOST = "https://dummy_cis2_auth.com"
+process.env.CIS2_OIDCJWKS_ENDPOINT = `${process.env.CIS2_OIDC_HOST}/.well-known/jwks.json`
+process.env.CIS2_USER_INFO_ENDPOINT = `${process.env.CIS2_OIDC_HOST}/userinfo`
+process.env.CIS2_USER_POOL_IDP = "CIS2DummyPoolIdentityProvider"
+process.env.CIS2_IDP_TOKEN_PATH = `${process.env.CIS2_OIDC_HOST}/token`
+
+process.env.MOCK_OIDC_ISSUER = "valid_mock_iss"
+process.env.MOCK_OIDC_CLIENT_ID = "valid_mock_aud"
+process.env.MOCK_OIDC_HOST = "https://dummy_mock_auth.com"
+process.env.MOCK_OIDCJWKS_ENDPOINT = `${process.env.MOCK_OIDC_HOST}/.well-known/jwks.json`
+process.env.MOCK_USER_INFO_ENDPOINT = `${process.env.MOCK_OIDC_HOST}/userinfo`
+process.env.MOCK_USER_POOL_IDP = "MockDummyPoolIdentityProvider"
+process.env.MOCK_IDP_TOKEN_PATH = `${process.env.MOCK_OIDC_HOST}/token`
